Revert "hypothesis testing"
This reverts commit 8b1e063f231938992a4c4f156aca08edd029cce1.
