+++ energypylinear/chp.py
@@ -234,4 +234,4 @@ class Generator:
         status = self.optimizer.solve()
         print(status)
 
-        return epl.results.extract_results(interval_data, vars)
+        return epl.data.extract_results(interval_data, vars)
