+++ energypylinear/objectives.py
@@ -29,11 +29,10 @@ def price_objective(
         + spills[i].electric_generation_mwh * defaults.spill_objective_penalty
         + spills[i].high_temperature_generation_mwh * defaults.spill_objective_penalty
         + spills[i].electric_load_mwh * defaults.spill_objective_penalty
+        #  dumping heat has no penalty
         + spills[i].high_temperature_load_mwh
         + [
-            spill_ev.charge_mwh
-            * defaults.spill_objective_penalty
-            * interval_data.electricity_prices[i]
+            spill_ev.charge_mwh * defaults.spill_objective_penalty
             for spill_ev in spill_evs[i]
         ]
         + [
