+++ tests/assets/test_evs.py
@@ -1,6 +1,7 @@
 import numpy as np
 
 import energypylinear as epl
+from energypylinear.freq import Freq
 
 """
 TODO
